remove .env.test only if it exists

+++ epicevents/epicevents/runner.py
@@ -5,4 +5,5 @@ from django.test.runner import DiscoverRunner
 class MyTestRunner(DiscoverRunner):
     def teardown_test_environment(self, **kwargs):
         super(MyTestRunner, self).teardown_test_environment(**kwargs)
-        os.remove('.env.test')
+        if os.path.exists('.env.test'):
+            os.remove('.env.test')
